pyproject.toml
[project]
name = "local-kb"
version = "0.1.0"
description = "Personal Local Knowledge Base (RAG) system"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.30.0",
    "faiss-cpu>=1.7.4",
    "chromadb>=0.4.22",
    "sentence-transformers>=2.3.1",
    "torch>=2.1.0",
    "transformers>=4.36.0",
    "pypdf>=4.0.0",
    "beautifulsoup4>=4.12.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "tiktoken>=0.5.2",
    "numpy>=1.26.0",
    "pandas>=2.1.0",
    "ollama>=0.1.0",
    "tqdm>=4.66.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/kb"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "black>=24.0.0",
    "isort>=5.13.0",
    "mypy>=1.8.0",
]
